Add files via upload
n this script, I used the PIL (Python Imaging Library) module, specifically the Image class from the Pillow package, to handle image processing tasks. I also used the os module to access and loop through all files in a specified folder. The script defines a function called resize() that takes a width and height as input. It goes through each file in the given folder, checks if it is a file (not a folder), and then opens the image using Image.open(). Each image is resized to the specified dimensions using the .resize() method and then saved with a new filename prefixed by "resized-". The purpose of the script is to automate the resizing of all images in a folder to a fixed size.